Fix the iOS build: gyro is Float, and NSLock is noasync
Two compile errors from the feeder and storage changes, both only visible
on the macOS runner:

  * `CapturedFrame.gyroDelta` is `(Float, Float, Float)` and
    `bs_frame_in.gyro_*` are floats; the feeder's owned copy declared
    `SIMD3<Double>` and failed at both ends.
  * `NSLock.lock()` is `@available(*, noasync)` — holding one across a
    suspension point deadlocks — and the storage path took it inline
    inside the `Task` that does the write. Moved into `finishStore()`, a
    plain synchronous method that awaits nothing, which is the property
    that makes it safe. A failed ENCODE now goes through the same path as
    a failed write: same outcome for the user, so it should count the
    same way.

Also folds the track-completion stage (S8) and this session's backlog
corrections in with it — see the next commit's message for the numbers.

+++ ios/Sources/Capture/EngineFeeder.swift
@@ -44,7 +44,7 @@ final class EngineFeeder: @unchecked Sendable {
         var depthWidth: Int
         var depthHeight: Int
         var dfx: Double, dfy: Double, dcx: Double, dcy: Double
-        var gyro: SIMD3<Double>?
+        var gyro: SIMD3<Float>?
     }
 
     /// Width the engine tracks at. It downsamples anything wider than 1280 to
